kernel: sched: Fix 32-bit to 64-bit cast warning

+++ kernel/sched/core.c
@@ -6122,7 +6122,7 @@ int sched_unisolate_cpu_unlocked(int cpu)
 		stop_cpus(cpumask_of(cpu), do_unisolation_work_cpu_stop, 0);
 
 		/* Kick CPU to immediately do load balancing */
-		if (!test_and_set_bit(NOHZ_BALANCE_KICK, (unsigned long *)nohz_flags(cpu)))
+		if (!test_and_set_bit(NOHZ_BALANCE_KICK, (unsigned long *)(uintptr_t)nohz_flags(cpu)))
 			smp_send_reschedule(cpu);
 	}
